tests/wpt/tests/css/css-inline/baseline-source/baseline-source-valid.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-inline-3/#baseline-source">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_valid_value('baseline-source', 'auto');
test_valid_value('baseline-source', 'first');
test_valid_value('baseline-source', 'last');
</script>
